+++ src/realm/object-store/sync/app_utils.cpp
@@ -23,7 +23,7 @@ namespace app {
 
 util::Optional<AppError> AppUtils::check_for_errors(const Response& response)
 {
-    if (response.status == ResponseStatus::failure && response.error.has_value()) {
+    if (response.status == ResponseStatus::Failure && response.error) {
         return *response.error;
     }